Final cleanup and refactoring.
