+++ airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/check/CheckOperation.kt
@@ -5,11 +5,6 @@
 package io.airbyte.cdk.load.check
 
 import io.airbyte.cdk.Operation
-import io.airbyte.cdk.command.ConfigurationSpecification
-import io.airbyte.cdk.command.ConfigurationSpecificationSupplier
-import io.airbyte.cdk.load.command.DestinationConfiguration
-import io.airbyte.cdk.load.command.DestinationConfigurationFactory
-import io.airbyte.cdk.output.ExceptionHandler
 import io.airbyte.cdk.output.OutputConsumer
 import io.airbyte.protocol.models.v0.AirbyteConnectionStatus
 import io.airbyte.protocol.models.v0.AirbyteMessage
@@ -21,31 +16,13 @@ private val logger = KotlinLogging.logger {}
 
 @Singleton
 @Requires(property = Operation.PROPERTY, value = "check")
-@Requires(env = ["destination"])
-class CheckOperation<T : ConfigurationSpecification, C : DestinationConfiguration>(
-    val configJsonObjectSupplier: ConfigurationSpecificationSupplier<T>,
-    val configFactory: DestinationConfigurationFactory<T, C>,
-    val destinationChecker: DestinationChecker<C>,
-    private val exceptionHandler: ExceptionHandler,
+class CheckOperation(
+    private val destinationChecker: DestinationChecker,
     private val outputConsumer: OutputConsumer,
 ) : Operation {
     override fun execute() {
-        val pojo =
-            try {
-                configJsonObjectSupplier.get()
-            } catch (e: Exception) {
-                handleException(e)
-                return
-            }
-        val config =
-            try {
-                configFactory.make(pojo)
-            } catch (e: Exception) {
-                handleException(e)
-                return
-            }
         try {
-            destinationChecker.check(config)
+            destinationChecker.check()
             val successMessage =
                 AirbyteMessage()
                     .withType(AirbyteMessage.Type.CONNECTION_STATUS)
@@ -56,15 +33,9 @@ class CheckOperation<T : ConfigurationSpecification, C : DestinationConfiguratio
             outputConsumer.accept(successMessage)
         } catch (t: Throwable) {
             logger.warn(t) { "Caught throwable during CHECK" }
-            handleException(t)
+            throw t
         } finally {
-            destinationChecker.cleanup(config)
+            destinationChecker.cleanup()
         }
-    }
-
-    private fun handleException(t: Throwable) {
-        val (traceMessage, statusMessage) = exceptionHandler.handleCheckFailure(t)
-        outputConsumer.accept(traceMessage)
-        outputConsumer.accept(statusMessage)
     }
 }
