+++ airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/command/DestinationStreamFactory.kt
@@ -59,31 +59,7 @@ class DestinationStreamFactory(
             minimumGenerationId = stream.minimumGenerationId,
             syncId = stream.syncId,
             schema = airbyteSchemaType,
-            isFileBased = stream.stream.isFileBased ?: false,
-            includeFiles = stream.includeFiles ?: false,
-            destinationObjectName = stream.destinationObjectName,
-            matchingKey =
-                stream.destinationObjectName?.let {
-                    fromCompositeNestedKeyToCompositeKey(stream.primaryKey)
-                },
             tableSchema = tableSchema,
         )
-    }
-
-    private fun fromCompositeNestedKeyToCompositeKey(
-        compositeNestedKey: List<List<String>>
-    ): List<String> {
-        if (compositeNestedKey.any { it.size > 1 }) {
-            throw IllegalArgumentException(
-                "Nested keys are not supported for matching keys. Key was $compositeNestedKey",
-            )
-        }
-        if (compositeNestedKey.any { it.isEmpty() }) {
-            throw IllegalArgumentException(
-                "Parts of the composite key need to have at least one element. Key was $compositeNestedKey",
-            )
-        }
-
-        return compositeNestedKey.map { it[0] }.toList()
     }
 }
