+++ airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/message/DestinationRecordSource.kt
@@ -26,7 +26,6 @@ import io.airbyte.protocol.protobuf.AirbyteMessage.AirbyteMessageProtobuf
 sealed interface DestinationRecordSource {
     val emittedAtMs: Long
     val sourceMeta: Meta
-    val fileReference: FileReference?
     fun asJsonRecord(orderedSchema: Array<FieldAccessor>): JsonNode
     fun asAirbyteValueProxy(): AirbyteValueProxy
 }
@@ -48,9 +47,6 @@ value class DestinationRecordJsonSource(val source: AirbyteMessage) : Destinatio
                     }
                         ?: emptyList()
             )
-
-    override val fileReference: FileReference?
-        get() = source.record.fileReference?.let { FileReference.fromProtocol(it) }
 
     override fun asJsonRecord(orderedSchema: Array<FieldAccessor>): JsonNode = source.record.data
 
@@ -76,9 +72,6 @@ value class DestinationRecordProtobufSource(val source: AirbyteMessageProtobuf) 
                     }
                         ?: emptyList()
             )
-
-    override val fileReference: FileReference?
-        get() = null
 
     override fun asJsonRecord(orderedSchema: Array<FieldAccessor>): JsonNode =
         asAirbyteValueProxy().asJson(orderedSchema)
