+++ airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/table/directload/DirectLoadTableStreamLoader.kt
@@ -8,7 +8,6 @@ import io.airbyte.cdk.load.command.DestinationStream
 import io.airbyte.cdk.load.component.TableOperationsClient
 import io.airbyte.cdk.load.component.TableSchemaEvolutionClient
 import io.airbyte.cdk.load.schema.model.TableName
-import io.airbyte.cdk.load.state.StreamProcessingFailed
 import io.airbyte.cdk.load.table.ColumnNameMapping
 import io.airbyte.cdk.load.table.TempTableNameGenerator
 import io.airbyte.cdk.load.write.StreamLoader
@@ -64,7 +63,7 @@ class DirectLoadTableAppendStreamLoader(
         streamStateStore.put(stream.mappedDescriptor, DirectLoadTableExecutionConfig(realTableName))
     }
 
-    override suspend fun close(hadNonzeroRecords: Boolean, streamFailure: StreamProcessingFailed?) {
+    override suspend fun teardown(completedSuccessfully: Boolean) {
         // do nothing
     }
 }
@@ -107,7 +106,7 @@ class DirectLoadTableDedupStreamLoader(
         streamStateStore.put(stream.mappedDescriptor, DirectLoadTableExecutionConfig(tempTableName))
     }
 
-    override suspend fun close(hadNonzeroRecords: Boolean, streamFailure: StreamProcessingFailed?) {
+    override suspend fun teardown(completedSuccessfully: Boolean) {
         if (initialStatus.realTable != null) {
             schemaEvolutionClient.ensureSchemaMatches(stream, realTableName, columnNameMapping)
         } else {
@@ -234,8 +233,8 @@ class DirectLoadTableAppendTruncateStreamLoader(
         )
     }
 
-    override suspend fun close(hadNonzeroRecords: Boolean, streamFailure: StreamProcessingFailed?) {
-        if (streamFailure == null && isWritingToTemporaryTable) {
+    override suspend fun teardown(completedSuccessfully: Boolean) {
+        if (completedSuccessfully && isWritingToTemporaryTable) {
             logger.info {
                 "Overwriting ${tempTableName.toPrettyString()} with ${realTableName.toPrettyString()} for stream ${stream.mappedDescriptor.toPrettyString()}"
             }
@@ -329,8 +328,8 @@ class DirectLoadTableDedupTruncateStreamLoader(
         streamStateStore.put(stream.mappedDescriptor, DirectLoadTableExecutionConfig(tempTableName))
     }
 
-    override suspend fun close(hadNonzeroRecords: Boolean, streamFailure: StreamProcessingFailed?) {
-        if (streamFailure == null) {
+    override suspend fun teardown(completedSuccessfully: Boolean) {
+        if (completedSuccessfully) {
             if (shouldCheckRealTableGeneration && shouldUpsertDirectly()) {
                 // Direct upsert path for simpler cases
                 logger.info {
