+++ airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/write/DestinationWriter.kt
@@ -5,9 +5,6 @@
 package io.airbyte.cdk.load.write
 
 import io.airbyte.cdk.load.command.DestinationStream
-import io.airbyte.cdk.load.state.DestinationFailure
-import io.micronaut.context.annotation.Secondary
-import jakarta.inject.Singleton
 
 /**
  * Implementor interface. Every Destination must extend this and at least provide an implementation
@@ -21,20 +18,5 @@ interface DestinationWriter {
     fun createStreamLoader(stream: DestinationStream): StreamLoader
 
     // Called once at the end of the job, unconditionally.
-    // NOTE: we don't pass Success here, because it depends on this completing successfully.
-    suspend fun teardown(destinationFailure: DestinationFailure? = null) {}
-}
-
-@Singleton
-@Secondary
-class DefaultDestinationWriter : DestinationWriter {
-    init {
-        throw NotImplementedError(
-            "DestinationWrite not implemented. Please create a custom @Singleton implementation."
-        )
-    }
-
-    override fun createStreamLoader(stream: DestinationStream): StreamLoader {
-        throw NotImplementedError()
-    }
+    suspend fun teardown(hadFailure: Boolean = false) {}
 }
