--- airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/write/LoadStrategy.kt
@@ -1,10 +1,0 @@
-/*
- * Copyright (c) 2026 Airbyte, Inc., all rights reserved.
- */
-
-package io.airbyte.cdk.load.write
-
-interface LoadStrategy {
-    val inputPartitions: Int
-        get() = 1
-}
