+++ airbyte-cdk/bulk/core/load/src/test/kotlin/io/airbyte/cdk/load/command/DestinationCatalogTest.kt
@@ -34,20 +34,17 @@ class DestinationCatalogTest {
                         .withMinimumGenerationId(34)
                         .withGenerationId(56)
                         .withDestinationSyncMode(DestinationSyncMode.APPEND)
-                        .withIncludeFiles(false)
                         .withStream(
                             AirbyteStream()
                                 .withJsonSchema("""{"type": "object"}""".deserializeToNode())
                                 .withNamespace("namespace1")
                                 .withName("name1")
-                                .withIsFileBased(false)
                         ),
                     ConfiguredAirbyteStream()
                         .withSyncId(12)
                         .withMinimumGenerationId(34)
                         .withGenerationId(56)
                         .withDestinationSyncMode(DestinationSyncMode.APPEND_DEDUP)
-                        .withIncludeFiles(true)
                         .withStream(
                             AirbyteStream()
                                 .withJsonSchema(
@@ -55,7 +52,6 @@ class DestinationCatalogTest {
                                 )
                                 .withNamespace("namespace2")
                                 .withName("name2")
-                                .withIsFileBased(true)
                         )
                         .withPrimaryKey(listOf(listOf("id1"), listOf("id2")))
                         .withCursorField(listOf("cursor")),
@@ -64,20 +60,17 @@ class DestinationCatalogTest {
                         .withMinimumGenerationId(34)
                         .withGenerationId(56)
                         .withDestinationSyncMode(DestinationSyncMode.OVERWRITE)
-                        .withIncludeFiles(false)
                         .withStream(
                             AirbyteStream()
                                 .withJsonSchema("""{"type": "object"}""".deserializeToNode())
                                 .withNamespace("namespace3")
                                 .withName("name3")
-                                .withIsFileBased(false)
                         ),
                     ConfiguredAirbyteStream()
                         .withSyncId(12)
                         .withMinimumGenerationId(34)
                         .withGenerationId(56)
                         .withDestinationSyncMode(DestinationSyncMode.APPEND_DEDUP)
-                        .withIncludeFiles(false)
                         .withStream(
                             AirbyteStream()
                                 .withJsonSchema(
@@ -85,7 +78,6 @@ class DestinationCatalogTest {
                                 )
                                 .withNamespace("namespace4")
                                 .withName("name4")
-                                .withIsFileBased(true)
                         )
                         .withPrimaryKey(listOf(listOf("id1"), listOf("id2")))
                         .withCursorField(listOf("cursor")),
@@ -102,7 +94,6 @@ class DestinationCatalogTest {
                 generationId = 1,
                 minimumGenerationId = 0,
                 syncId = 1,
-                includeFiles = false,
                 schema =
                     ObjectType(
                         properties =
@@ -155,7 +146,6 @@ class DestinationCatalogTest {
                             generationId = 1,
                             minimumGenerationId = 0,
                             syncId = 1,
-                            includeFiles = false,
                             schema = ObjectType(linkedMapOf()),
                             namespaceMapper = NamespaceMapper(),
                             tableSchema =
@@ -178,7 +168,6 @@ class DestinationCatalogTest {
                             generationId = 1,
                             minimumGenerationId = 0,
                             syncId = 1,
-                            includeFiles = false,
                             schema = ObjectType(linkedMapOf()),
                             namespaceMapper = NamespaceMapper(),
                             tableSchema =
@@ -214,7 +203,6 @@ class DestinationCatalogTest {
                             generationId = 1,
                             minimumGenerationId = 0,
                             syncId = 1,
-                            includeFiles = false,
                             schema = ObjectType(linkedMapOf()),
                             namespaceMapper = NamespaceMapper(),
                             tableSchema =
@@ -260,7 +248,6 @@ class DestinationCatalogTest {
                             generationId = 1,
                             minimumGenerationId = 0,
                             syncId = 1,
-                            includeFiles = false,
                             schema =
                                 ObjectType(
                                     linkedMapOf("id" to FieldType(IntegerType, nullable = true))
